Converted Slate Tag Webpack Plugin to TS
