+++ packages/slate-tag-webpack-plugin/tsconfig.json
@@ -1,0 +1,9 @@
+{
+  "extends": "../../tsconfig.json",
+  "compilerOptions": {
+    "outDir": "./dist/"
+  },
+  "include": [
+    "src/**/*"
+  ]
+}
